# AI-Powered-One-Stop-Travel-Solution/backend/openapi.yaml
openapi: 3.0.0
info:
  title: Hyper-Local AI Travel Marketplace API
  version: 1.0.0
  description: Backend REST API for the Hyper-Local AI Travel Marketplace
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/events:
    post:
      summary: Create event
      tags:
        - Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agencyId
                - title
                - location
                - startDate
                - price
                - totalSeats
              properties:
                agencyId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                location:
                  type: object
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
                price:
                  type: number
                totalSeats:
                  type: integer
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Event created
  /api/events/{id}:
    get:
      summary: Get event by ID
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event details
...